refactor: 에러 페이지에 대해 ErrorPage 하나로 사용하도록 변경
