+++ client/src/Router.tsx
@@ -3,7 +3,7 @@ import GlobalStyle, { MainLayout } from '@styles/GlobalStyle';
 import HomePage from '@pages/HomePage';
 import InitRoomPage from '@pages/InitRoomPage';
 import MainPage from '@pages/MainPage';
-import NotFoundPage from '@pages/NotFoundPage';
+import ErrorPage from '@pages/ErrorPage';
 
 function Router() {
   return (
@@ -15,7 +15,7 @@ function Router() {
           <Route path="/" element={<HomePage />} />
           <Route path="/init-room" element={<InitRoomPage />} />
           <Route path="/room/:roomCode" element={<MainPage />} />
-          <Route path="*" element={<NotFoundPage />} />
+          <Route path="*" element={<ErrorPage reason="존재하지 않는 페이지" />} />
         </Routes>
       </MainLayout>
     </BrowserRouter>
