+++ client/src/pages/ErrorPage/styles.tsx
@@ -1,4 +1,15 @@
 import styled from 'styled-components';
+import * as palette from '@styles/Variables';
+
+export const ErrorPageLayout = styled.div`
+  background-color: ${palette.PRIMARY};
+  width: 100%;
+  height: 100%;
+  display: flex;
+  justify-content: center;
+  align-items: center;
+  position: relative;
+`;
 
 export const ErrorBox = styled.div`
   display: flex;
