--- client/src/pages/NotFoundPage/styles.tsx
@@ -1,12 +1,0 @@
-import styled from 'styled-components';
-import * as palette from '@styles/Variables';
-
-export const NotFoundPageLayout = styled.div`
-  background-color: ${palette.PRIMARY};
-  width: 100%;
-  height: 100%;
-  display: flex;
-  justify-content: center;
-  align-items: center;
-  position: relative;
-`;
